Add module util
Util bundles different functions and structs that are used in the client. This makes the client module easier to maintain because it really only defines the main functionality of the client.
All tests still pass.
