+++ meteomatics/src/lib.rs
@@ -44,6 +44,7 @@
 pub mod errors;
 pub mod client;
 pub mod location;
+pub mod util;
 pub use client::APIClient;
 
 // TODO: check where to pass references and where to pass ownership
